#159428917 implement feedback given

+++ server/controller/entries.js
@@ -1,74 +1,80 @@
 import pool from '../database/connectDatabase';
-import InputValidator from '../utilities/inputvalidators';
 
+/**
+ * @export EntriesController
+ * @class EntriesController
+ */
 class EntriesController {
   /**
-   * A controller to fetch all entries from the database.
-   */
+ * @function fetchEntries
+ * @param {*} req
+ * @param {*} res
+ * @returns {*} All specific user entries
+ */
   static fetchEntries(req, res) {
     // retrieve all entries from the database FOR NOW!!!.
-    const fetchEntryQuery = 'SELECT * FROM entries;';
-    pool.query(fetchEntryQuery, (err, result) => {
-      if (result.rowCount < 1) {
-        return res.status(404).json({
-          message: 'No entry',
+    const fetchEntryQuery = `SELECT * FROM entries WHERE userid = '${req.decoded.userId}';`;
+    try {
+      pool.query(fetchEntryQuery, (err, result) => {
+        if (result.rowCount < 1) {
+          return res.status(404).json({
+            message: 'No entry',
+          });
+        }
+        return res.status(200).json({
+          message: 'All entries by users',
+          entries: result.rows,
         });
-      }
-      return res.status(200).json({
-        message: 'All entries by users',
-        entries: result.rows,
       });
-    });
-  }
+    } catch (err) {
+      return res.status(500).json({
+        message: 'Something went wrong!',
+      });
+    }
+    return null;
+  } /**
+  * @function fetchEntries
+  * @param {*} req
+  * @param {*} res
+  * @returns {*} Email notification
+  */
 
-  /**
-   * A controller to fetch a specific entry
-   */
   static fetchSingleEntry(req, res) {
     const { entriesId } = req.params;
-    // Searches the particular entry from the database
-    const findEntryQuery = `SELECT * FROM entries WHERE entriesId = '${entriesId}';`;
-    pool.query(findEntryQuery, (err, result) => {
-      if (result.rowCount < 1) {
-        return res.status(404).json({
-          message: 'No such entry',
-        });
-      }
-      return res.status(200).json({
-        message: 'Entry found',
-        entry: result.rows[0],
-      });
-    });
-  }
+    const findEntryQuery = `SELECT * FROM entries WHERE entriesId = '${entriesId}' AND userid = '${req.decoded.userId}';`;
 
+    try {
+      pool.query(findEntryQuery, (err, result) => {
+        if (result.rowCount < 1) {
+          return res.status(404).json({
+            message: 'No such entry',
+          });
+        }
+        return res.status(200).json({
+          message: 'Entry found',
+          entry: result.rows[0],
+        });
+      });
+    } catch (err) {
+      return res.status(500).json({
+        message: 'Something went wrong!',
+      });
+    }
+    return null;
+  }
   /**
- * A controller to add an new entry to the database
- *    Create an object then push to the database.
+ * @function createEntry
+ * @param {*} req
+ * @param {*} res
+ * @returns {*} the created entry
  */
-  // eslint-disable-next-line
+
   static createEntry(req, res) {
-    // Get parameters from the req.body
     const {
-      entryTitle, entry, visibility, userId,
+      entryTitle, entry, visibility,
     } = req.body;
-    // Validation happens here
-    if (InputValidator.validateEntryTitle(entryTitle) === false) {
-      return res.status(400).json({
-        message: 'Entry title required!',
-      });
-    }
-    if (InputValidator.validateEntry(entry) === false) {
-      return res.status(400).json({
-        message: 'Pls, enter an entry',
-      });
-    }
-    if (InputValidator.validateEntryVisibility(visibility) === false) {
-      return res.status(400).json({
-        message: 'Entry visibility is required!',
-      });
-    }
-    // Gets userId from the token
-    const createEntryQuery = `INSERT INTO entries (entry, entryTitle, visibility, userId) VALUES ('${entry}', '${entryTitle}', '${visibility}', '${userId}');`;
+
+    const createEntryQuery = `INSERT INTO entries (entry, entryTitle, visibility, userId) VALUES ('${entry}', '${entryTitle}', '${visibility}', '${req.decoded.userId}');`;
     pool.query(createEntryQuery, (err, result) => {
       if (result.rowCount < 1) {
         return res.status(400).json({
@@ -83,35 +89,19 @@ class EntriesController {
   }
 
   /**
-   * A controller to update an entry
-   */
+ * @function modifyEntry
+ * @param {*} req
+ * @param {*} res
+ * @returns {*} the entry modified
+ */
   static modifyEntry(req, res) {
     // Collects the entriesId
     const { entriesId } = req.params;
-    // Get parameters from the req.body
     const {
       entryTitle, entry, visibility,
     } = req.body;
-
-    // Validate entries
-    if (InputValidator.validateEntryTitle(entryTitle) === false) {
-      return res.status(400).json({
-        message: 'Entry title required!',
-      });
-    }
-    if (InputValidator.validateEntry(entry) === false) {
-      return res.status(400).json({
-        message: 'Pls, enter an entry',
-      });
-    }
-    if (InputValidator.validateEntryVisibility(visibility) === false) {
-      return res.status(400).json({
-        message: 'Entry visibility is required!',
-      });
-    }
-
     // Update SQL query
-    const updateEntryQuery = `UPDATE entries SET entryTitle = '${entryTitle}', entry = '${entry}', visibility = '${visibility}' WHERE entriesId = '${parseInt(entriesId, 10)}';`;
+    const updateEntryQuery = `UPDATE entries SET entryTitle = '${entryTitle}', entry = '${entry}', visibility = '${visibility}' WHERE entriesId = '${parseInt(entriesId, 10)}' AND userId = '${req.decoded.userId}' RETURNING *;`;
     pool.query(updateEntryQuery, (err, result) => {
       if (result.rowCount < 0) {
         return res.status(404).json({
@@ -120,24 +110,23 @@ class EntriesController {
       }
       return res.status(200).json({
         message: 'successfully updated',
-        entry: result.rows,
+        entry: result.rows[0],
       });
     });
     return null;
   }
 
   /**
-   * A controller to delete an entry
-   */
+ * @function deleteEntry
+ * @param {*} req
+ * @param {*} res
+ * @returns {*} the deleted entry
+ */
   static deleteEntry(req, res) {
-    // Collects the entriesId
     const { entriesId } = req.params;
-
-    // Delete entry from entries where entryId is something
-    const deleteEntryQuery = `DELETE FROM entries WHERE entriesId = ${entriesId};`;
+    const deleteEntryQuery = `DELETE FROM entries WHERE entriesId = '${entriesId}' AND userId = '${req.decoded.userId}' RETURNING *;`;
 
     pool.query(deleteEntryQuery, (err, result) => {
-      // Does not show content but response is deleted.
       if (result.rowCount < 1) {
         return res.status(404).json({
           message: 'User not found!!',
@@ -145,20 +134,33 @@ class EntriesController {
       }
       return res.status(200).json({
         message: 'User Deleted!!',
+        userDeleted: result.rows[0],
       });
     });
 
     //
   }
 
+  /**
+ * @function countEntries
+ * @param {*} req
+ * @param {*} res
+ * @returns {*} the total number of entries for a specific user
+ */
   static countEntries(req, res) {
-    // Get userId from token
-    const countQuery = 'SELECT COUNT(entriesId) FROM entries';
+    const countQuery = `SELECT COUNT(entriesId) FROM entries WHERE userId = '${req.decoded.userId}'`;
     pool.query(countQuery, (err, result) => {
-      return res.status(200).json({
+      if (err) {
+        return res.status(500).json({
+          message: 'Internal Server error',
+        });
+      }
+      res.status(200).json({
         message: 'Entries count',
-        count: result.rows[0],
+        count: result.rows[0].count,
+
       });
+      return null;
     });
     return null;
   }
